@@ -20,9 +20,9 @@ func (r *exampleQuery) GetExampleByID(id int64, organizationID int64) (*Example,
 	var example Example
 	var err error
 	if organizationID != 0 {
-		err = r.conn.Get(&example, "SELECT * FROM examples WHERE id = ? AND organization_id = ?", id, organizationID)
+		err = r.conn.Get(&example, "SELECT * FROM examples WHERE id = ? AND organization_id = ? AND status > 0", id, organizationID)
 	} else {
-		err = r.conn.Get(&example, "SELECT * FROM examples WHERE id = ? ", id)
+		err = r.conn.Get(&example, "SELECT * FROM examples WHERE id = ? AND status > 0", id)
 	}
 	if err != nil {
 		return nil, err
@@ -31,7 +31,7 @@ func (r *exampleQuery) GetExampleByID(id int64, organizationID int64) (*Example,
 }
 
 func (r *exampleQuery) GetExampleCount(filter ExampleFilter) (int, error) {
-	where, args := []string{"1 = 1"}, []interface{}{}
+	where, args := []string{"status > 0"}, []interface{}{}
 	if v := filter.Name; v != "" {
 		where, args = append(where, "name like ?"), append(args, "%"+v+"%")
 	}
@@ -73,7 +73,7 @@ func (r *exampleQuery) GetExampleCount(filter ExampleFilter) (int, error) {
 }
 
 func (r *exampleQuery) GetExampleList(filter ExampleFilter) (*[]ExampleListResponse, error) {
-	where, args := []string{"1 = 1"}, []interface{}{}
+	where, args := []string{"status > 0"}, []interface{}{}
 	if v := filter.Name; v != "" {
 		where, args = append(where, "e.name like ?"), append(args, "%"+v+"%")
 	}
